requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
decorator==5.1.1
dj-static==0.0.6
Django==4.2
django-bootstrap-breadcrumbs==0.9.2
django-bootstrap5==23.1
django-grappelli==3.0.6
djangorestframework==3.14.0
djlint==1.25.0
EditorConfig==0.12.3
executing==1.2.0
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
ipython==8.13.1
jedi==0.18.2
jsbeautifier==1.14.7
matplotlib-inline==0.1.6
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pytz==2023.3
PyYAML==6.0
regex==2023.5.4
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
static3==0.7.0
tomli==2.0.1
tqdm==4.65.0
traitlets==5.9.0
wcwidth==0.2.6
